fix(comp): cleans up DateSelector component and tests
